Adapt to restructure
