+++ frontend/src/lib/components/common/Page.svelte
@@ -50,7 +50,7 @@
       );
     }
     .page {
-      margin: 3rem 0rem 0rem 0rem;
+      margin: 0rem 0rem 0rem 0rem;
       padding: 0;
     }
   </style>
